test-runner: Simplify disk test
Per https://github.com/rust-osdev/uefi-rs/issues/553, we've made the test runner
always expect to be running in the special QEMU env, so we don't need to have a
separate "works with any env" disk test from the more detailed "known disk"
test. So, move `known_disk.rs` -> `mod.rs`.
